DesignTK_Final/requirements.txt
openai
streamlit
python-dotenv
scikit-learn
plotly